+++ R/AppenderDbi.R
@@ -8,7 +8,7 @@
 #' Log to a database table with any **DBI** compatible backend. Please be
 #' aware that AppenderDbi does *not* support case sensitive / quoted column
 #' names, and you advised to only use all-lowercase names for
-#' custom fields (see `...` argument of [LogEvent]).
+#' custom fields (see `...` argument of [lgr::LogEvent]).
 #' When appending to a database table all LogEvent values for which a column
 #' exists in the target table will be appended, all others are ignored.
 #'
@@ -76,7 +76,7 @@ AppenderDbi <- R6::R6Class(
   public = list(
 
     #' @param conn,table see section *Fields*
-    #' @param threshold,flush_threshold,layout,buffer_size see [AppenderBuffer]
+    #' @param threshold,flush_threshold,layout,buffer_size see [lgr::AppenderBuffer]
     initialize = function(
       conn,
       table,